Refactor to add authorization
