+++ src/middleware/packages/connector/package.json
@@ -11,7 +11,9 @@
     "jsonwebtoken": "^8.5.1",
     "passport-cas2": "0.0.11",
     "node-jose": "^1.1.3",
-    "openid-client": "^3.11.0"
+    "openid-client": "^3.11.0",
+    "moleculer": "^0.13.11",
+    "moleculer-web": "^0.8.5"
   },
   "author": "",
   "license": "Apache-2.0"
